Adding Auth
- Adding Missing DTOs
- Removing innecessary models
- Adding Profiles
- Finishing Assignment, User, Course and AuthControllers
- Fixing Issues
- Cleaning solution

+++ TeacherControl/TeacherControl.Core/Models/Course/CourseUserCredit.cs
@@ -7,9 +7,9 @@ namespace TeacherControl.Core.Models
 {
     public class CourseUserCredit : IModificationAudit
     {
-        //Add a status here for the user credits?
+        //TBD: Add a status here for the user credits?
         public int Id { get; set; }
-        public float Credits { get; set; }
+        public double Credits { get; set; }
         public int StudentId { get; set; }
         public virtual User Student { get; set; }
         public int CourseId { get; set; }
